fix black+friend

+++ DatingWebsiteApp/App.WebAPI/Controllers/BlackListController.cs
@@ -40,7 +40,7 @@ namespace App.WebAPI.Controllers
         public async Task<IActionResult> GetBlackListWithMe()
         {
             var user_id = User.Claims.First(c => c.Type == "UserID").Value;
-            var black_list = await _blackListService.GetMyBlackListAsync(user_id);
+            var black_list = await _blackListService.GetBlackListWithMeAsync(user_id);
             if (black_list == null)
             {
                 return NotFound(new { message = "User not found by id." });
@@ -48,24 +48,34 @@ namespace App.WebAPI.Controllers
             return Ok(black_list);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet]
+        [Route("{id}")]
         [Authorize]
         public async Task<IActionResult> CreateBlack(string id)  //bad_guy_id
         {
             var user_id = User.Claims.First(c => c.Type == "UserID").Value;
+            if (user_id==id)
+            {
+                return BadRequest(new { message = "You cannot (user_id == bad_guy_id)." });
+            }
             var new_bad = await _blackListService.AddToBlackListAsync(user_id, id);
             if (new_bad == null)
             {
-                return NotFound(new { message = "User not found by id." });
+                return NotFound(new { message = "User not found by id (or user_in_black_list already exist)." });
             }
             return Ok(new_bad);
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete]
+        [Route("{id}")]
         [Authorize]
         public async Task<IActionResult> DeleteBlack(string id)  //bad_guy_id
         {
             var user_id = User.Claims.First(c => c.Type == "UserID").Value;
+            if (user_id == id)
+            {
+                return BadRequest(new { message = "You cannot (user_id == bad_guy_id)." });
+            }
             var new_good = await _blackListService.DeleteFromBlackListAsync(user_id, id);
             if (new_good == null)
             {
